aider: Added option to specify virtual environment location and activated it before running Python scripts in `new_script.sh` and `new_script.bat`.
# Aider chat conversation:

USER: analyze new_script files for cleanliness and functionality.

USER: Ensure that the python virtual environment is activated prior to running any functions. Assume that this is a venv virtual environment in the parent folder of the directory the script is being run in, and allow user to configure a different location for this virtual environment as a script option.

ASSISTANT: To ensure that the Python virtual environment is activated prior to running any functions and to allow the user to configure a different location for this virtual environment, we need to:

1. Add an option for the user to specify the virtual environment location.
2. Activate the virtual environment before running any Python scripts.

Here are the changes for both `new_script.sh` and `new_script.bat`:

### For `new_script.sh`:

1. Add an option to specify the virtual environment location.
2. Activate the virtual environment before running any Python scripts.

new_script.sh
```sh
<<<<<<< SEARCH
while true; do
=======
VENV_PATH="../venv"
if [ ! -z "$1" ]; then
    VENV_PATH="$1"
fi

source "$VENV_PATH/bin/activate"

while true; do
>>>>>>> REPLACE
```

### For `new_script.bat`:

1. Add an option to specify the virtual environment location.
2. Activate the virtual environment before running any Python scripts.

new_script.bat
```batch
<<<<<<< SEARCH
@echo off
=======
@echo off
setlocal enabledelayedexpansion

set "VENV_PATH=..\venv"
if not "%1"=="" set "VENV_PATH=%1"

call "%VENV_PATH%\Scripts\activate"
>>>>>>> REPLACE
```

These changes will ensure that the virtual environment is activated before any Python scripts are executed, and they allow the user to specify a different location for the virtual environment if needed.